SGD removed, SBGD with batch_size set to 1 is the same.
I think storing the layers as a linked list hinting at a graph structure is more appropriate than a sequential model, though the sequential model could technically be capable of residual connections if thats all that's needed.

an n head model can be trained by making the grads function return the final delta, this can be summed from n heads and fed into the main model as the error.